*: refactor the private session
