+++ pkg/executor/test/seqtest/seq_executor_test.go
@@ -43,6 +43,7 @@ import (
 	"github.com/pingcap/tidb/pkg/parser/mysql"
 	plannercore "github.com/pingcap/tidb/pkg/planner/core"
 	"github.com/pingcap/tidb/pkg/session"
+	"github.com/pingcap/tidb/pkg/session/internal"
 	"github.com/pingcap/tidb/pkg/sessionctx/variable"
 	"github.com/pingcap/tidb/pkg/store/copr"
 	"github.com/pingcap/tidb/pkg/store/mockstore"
@@ -1324,7 +1325,7 @@ func TestAutoRandIDRetry(t *testing.T) {
 	tk.MustExec("insert into t values (),()")
 	tk.MustExec("insert into t values ()")
 
-	session.ResetMockAutoRandIDRetryCount(5)
+	internal.ResetMockAutoRandIDRetryCount(5)
 	fpName := "github.com/pingcap/tidb/pkg/session/mockCommitRetryForAutoRandID"
 	require.NoError(t, failpoint.Enable(fpName, `return(true)`))
 	tk.MustExec("commit")
@@ -1333,7 +1334,7 @@ func TestAutoRandIDRetry(t *testing.T) {
 	maskedHandles := extractMaskedOrderedHandles()
 	require.Equal(t, []int64{1, 2, 3, 4, 5}, maskedHandles)
 
-	session.ResetMockAutoRandIDRetryCount(11)
+	internal.ResetMockAutoRandIDRetryCount(11)
 	tk.MustExec("begin")
 	tk.MustExec("insert into t values ()")
 	require.NoError(t, failpoint.Enable(fpName, `return(true)`))
